generalize checkpointing
